MAILBOX-318 Merge SetMailboxes stepdefs together
